external changes in the errors will be reflected in the internal implementation
