+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
             'minSdk'       : 16,
             'compileSdk'   : 29,
             'targetSdk'    : 29,
-            'kotlin'       : '1.3.61',
+            'kotlin'       : '1.3.70',
             'androidPlugin': '3.6.1',
             'androidArch'  : '1.1.1',
             'rxJava'       : '2.2.2',
